requirements.txt
aws-cdk-lib==2.55.1
constructs>=10.0.0,<11.0.0
boto3
requests
aws-cdk.aws-neptune-alpha
aws-cdk.aws-apigatewayv2-alpha
aws-cdk.aws-apigatewayv2-integrations-alpha

